refactor(cjson,onvif,network): hide some details of implementation, and fix some typos
We remove unessential definitions in the header files, so the header files installed can be shrunk.
It influences onvif_custom(we delete it because it contains only one source file) and
network(transfer some definitions). Notice that this process will involve all submodules of the
oranization, and this commit marks a start. We will release a BREAKING CHANGES following, but in the
theory, it would work the same as the previous one. However, we don't ensure that it works
compatibly, so please make a full test before upgrading to this patch. By the way, we fix a tiny
typo in cjson CMakelists, so that it looks normal.

BREAKING CHANGE: As the commit message says, we release BREAKING CHANGES in case that ner changes is
not compatible with the previous release. This message serves as a reminder of paying close
attention to potential risk.